requirements.txt
faster-whisper
sounddevice
soundfile
numpy
pynput
Pillow
pystray
